# app/swagger.yaml
openapi: 3.0.0
info:
  title: ODFDS BackEnd API
  description: ODFDS BackEnd API
  version: 1.0.0
servers:
  - url: http://164.92.68.162:3000
    description: Remote Development Server
  - url: http://127.0.0.1:3000
    description: Local Development Server
...
paths:
...
  /driver:
    post:
      description: Sign up a new driver accout
      tags:
      - Driver
      summary: Sign up a new driver accout
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverSignUpRequest'
      responses:
        '200':
          description: Sign up succeed, restaurant account data inserted to the database
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/GeneralResponse'
              example:
                code: 200
                data: Restaurant account created successfully
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralInvalidResponse'
              examples:
                Email verification code expired:
                  value: 
                    code: 400
                    data:
                      field: code
                      value: 123456
                      message: Email verification code is expired, please try to request a new one
                Wrong Email verification code:
                  value: 
                    code: 400
                    data:
                      field: code
                      value: 123456
                      message: Email verification code is wrong, please try again
                Email address already taken:
                  value:
                    code: 400
                    data:
                      field: email
                      value: steven.panxd@gmail.com
                      message: Email is already taken, please try another email address
    delete:
      description: Delete a driver account (dev use)
      tags:
      - Driver
      summary: Delete a driver account (dev use)
      parameters:
      - in: query
        name: email
        schema:
          type: string
        required: true
      responses:
        '200':
          description: Succeed
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/GeneralResponse'
              example:
                code: 200
                data: Driver account deleted
        '400':
          description: Fail
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/GeneralInvalidResponse'
              examples:
                Invalid email address format:
                  value:
                    code: 400
                    data:
                      field: email
                      value: stevenxpan@yahoo.com
                      message: Driver account does not exist
...
components:
  schemas:
    GeneralResponse:
      title: General valid response body
      type: object
      properties:
        code:
          type: number
          description: HTTP status code
          example: 200
        data:
          type: string
          description: response data or succeed message
          example: succeed
    GeneralInvalidResponse:
      title: General invalid parameter response body
      type: object
      properties:
        code:
          type: number
          description: HTTP status code
          example: 400
        data:
          type: object
          properties:
            field:
              type: string
              description: error field
            value:
              type: string
              description: backend received error value
            message:
              type: string
              description: error message
          example:
            field: email
            value: abc.com
            message: Invalid email address
...
    DriverSignUpRequest:
      title: Driver account sign up request body
      type: object
      properties:
        email:
          type: string
          description: email address, unique
          example: steven.panxd@gmail.com
        code:
          type: string
          description: email verification code
          example: 123456
        password:
          type: string
          description: password, much be length >= 6, at least 1 uppercase letter, 1 lowercase letter, and 1 symbol
          example: A1b2c.
        phone:
          type: string
          description: phone number, unique
          example: 4081234567
        driverLicenseNumber:
          type: string
          description: driver license number
          example: Y1234567
        driverLicenseImage:
          type: string
          description: driver license image url
          example: https://www.google.com/images/branding/googlelogo/1x/googlelogo_color_272x92dp.png
        firstName:
          type: string
          description: driver's first name
          example: Steven
        lastName:
          type: string
          description: driver's last name
          example: Pan
        middleName:
          type: string
          description: driver's middle name
          example: null
        bankAccountNumber:
          type: string
          description: driver's bank account number
          example: 123456789
        bankRoutingNumber:
          type: string
          description: driver's bank routing number
          example: 123456789
      required:
        - email
        - code
        - password
        - phone
        - driverLicenseNumber
        - driverLicenseImage
        - firstName
        - lastName
        - bankAccountNumber
        - bankRoutingNumber
...